CFLAGS

+++ lexer/lexer_utils.c
@@ -6,7 +6,7 @@
 /*   By: rnishimo <rnishimo@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 16:53:25 by rnishimo          #+#    #+#             */
-/*   Updated: 2022/07/02 23:46:36 by rnishimo         ###   ########.fr       */
+/*   Updated: 2022/07/03 13:43:57 by rnishimo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ t_token	*skip(t_token *tok, t_token_kind kind, char *str)
 	if (tok->kind != kind)
 		syntax_error(tok->str, tok->len);
 	if (str != NULL
-		&& (tok->len != ft_strlen(str)
+		&& (tok->len != (long)ft_strlen(str)
 			|| ft_strncmp(tok->str, str, tok->len) != 0))
 		syntax_error(tok->str, tok->len);
 	return (tok->next);
@@ -28,7 +28,7 @@ bool	equal(t_token *tok, t_token_kind kind, char *str)
 	if (tok->kind != kind)
 		return (false);
 	if (str != NULL
-		&& (tok->len != ft_strlen(str)
+		&& (tok->len != (long)ft_strlen(str)
 			|| ft_strncmp(tok->str, str, tok->len) != 0))
 		return (false);
 	return (true);
